ajustes

+++ rna/treinamento/Treino.java
@@ -76,12 +76,13 @@ public class Treino{
             otimizador.atualizar(camadas);
          }
 
-         // if(e % 20 == 0){
+         // if(e % 25 == 0){
          //    System.out.println("Perda (" + e + "): " + (double)(perdaEpoca/entrada.length));
          // }
+
          //feedback de avanço da rede
          if(this.calcularHistorico){
-            this.historico = aux.adicionarPerda(this.historico, perdaEpoca/entrada.length);
+            this.historico = aux.adicionarPerda(this.historico, (double)(perdaEpoca/entrada.length));
          }
       }
    }
